Move monoid and semigroup functions to the instance
