+++ core/src/main/scala/cats/InvariantSemigroupal.scala
@@ -14,26 +14,15 @@ import simulacrum.typeclass
        def G = Apply[G]
      }
 
+  /**
+    * Gives a `Semigroup` instance if A itself has a `Semigroup` instance.
+    */
+  def semigroup[A](implicit A: Semigroup[A]): Semigroup[F[A]] =
+    new InvariantSemigroupalSemigroup[F, A](this, A)
 
 }
 
-object InvariantSemigroupal extends SemigroupalArityFunctions {
-  def semigroup[F[_], A](implicit f: InvariantSemigroupal[F], sg: Semigroup[A]): Semigroup[F[A]] =
-    new InvariantSemigroupalSemigroup[F, A](f, sg)
-
-  implicit def catsSemigroupalForMonoid: InvariantSemigroupal[Monoid] = new InvariantSemigroupal[Monoid] {
-    def product[A, B](fa: Monoid[A], fb: Monoid[B]): Monoid[(A, B)] = new Monoid[(A, B)] {
-      val empty = fa.empty -> fb.empty
-      def combine(x: (A, B), y: (A, B)): (A, B) = fa.combine(x._1, y._1) -> fb.combine(x._2, y._2)
-    }
-
-    def imap[A, B](fa: Monoid[A])(f: A => B)(g: B => A): Monoid[B] = new Monoid[B] {
-      def empty: B = f(fa.empty)
-
-      def combine(x: B, y: B): B = f(fa.combine(g(x), g(y)))
-    }
-  }
-}
+object InvariantSemigroupal extends SemigroupalArityFunctions
 
 private[cats] class InvariantSemigroupalSemigroup[F[_], A](f: InvariantSemigroupal[F], sg: Semigroup[A]) extends Semigroup[F[A]] {
   def combine(a: F[A], b: F[A]): F[A] =
