+++ tests/src/test/scala/cats/tests/MonoidSuite.scala
@@ -19,7 +19,6 @@ class MonoidSuite extends CatsSuite {
 object MonoidSuite {
   def summonInstance(): Unit = {
     Invariant[Monoid]
-    InvariantSemigroupal[Monoid]
     ()
   }
